Rename `Utils::BottleAPI` to `BottleAPI`
