+++ Library/Homebrew/bottle_api.rbi
@@ -1,0 +1,5 @@
+# typed: strict
+
+module BottleAPI
+  include Kernel
+end
