+++ Library/Homebrew/cmd/outdated.rb
@@ -6,6 +6,7 @@ require "keg"
 require "cli/parser"
 require "cask/cmd"
 require "cask/caskroom"
+require "bottle_api"
 
 module Homebrew
   extend T::Sig
@@ -100,7 +101,7 @@ module Homebrew
           elsif f.tap.present?
             f.pkg_version.to_s
           else
-            Utils::BottleAPI.latest_pkg_version(f.name)&.to_s || f.pkg_version.to_s
+            BottleAPI.latest_pkg_version(f.name)&.to_s || f.pkg_version.to_s
           end
 
           outdated_versions = outdated_kegs.group_by { |keg| Formulary.from_keg(keg).full_name }
