+++ Library/Homebrew/utils.rb
@@ -4,7 +4,6 @@
 require "time"
 
 require "utils/analytics"
-require "utils/bottle_api"
 require "utils/curl"
 require "utils/fork"
 require "utils/formatter"
