--- Library/Homebrew/utils/bottle_api.rbi
@@ -1,7 +1,0 @@
-# typed: strict
-
-module Utils
-  module BottleAPI
-    include Kernel
-  end
-end
